Feature/refactor exceptions

+++ src/albuns/albuns.controller.ts
@@ -1,4 +1,16 @@
-import { Body, Controller, Delete, Get, Param, Query, Post, Put, UploadedFile, UploadedFiles, UseInterceptors } from '@nestjs/common';
+import { 
+    Body, 
+    Controller,
+    Delete, 
+    Get,
+    Param,
+    Query,
+    Post,
+    Put,
+    UploadedFile,
+    UseInterceptors,
+    HttpException,
+    HttpStatus } from '@nestjs/common'
 import { FileInterceptor } from '@nestjs/platform-express'
 import { AlbunsService } from './albuns.service';
 import { AlbunsDTO } from './dtos/albuns.dto'
@@ -9,28 +21,43 @@ export class AlbunsController {
     }
 
     @Get(':id')
-    get(@Param('id') id: string) {
-        return this.service.findById(id);
+    async get(@Param('id') id: string) {
+        
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+
+        return await this.service.findById(id);
     }
     
     @Get()
-    getAll(@Query('page') page: number) {
-        return this.service.findAll(page);
+    async getAll(@Query('page') page: number) {
+        return await this.service.findAll(page);
     }
 
     @Post()
     @UseInterceptors(FileInterceptor('file'))
-    create(@Body() albuns: AlbunsDTO, @UploadedFile() file: Express.Multer.File) {
-        return this.service.create(albuns, file);
+    async create(@Body() albuns: AlbunsDTO, @UploadedFile() file: Express.Multer.File) {
+        return await this.service.create(albuns, file);
     }
 
     @Put(':id')
-    update(@Param('id') id: string, @Body() albuns: AlbunsDTO) {
-        return this.service.update(id, albuns);
+    async update(@Param('id') id: string, @Body() albuns: AlbunsDTO) {
+
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+        
+        return await this.service.update(id, albuns);
     }
 
     @Delete(':id')
-    remove(@Param('id') id: string,) {
-        return this.service.remove(id);
+    async remove(@Param('id') id: string,) {
+
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+
+        return await this.service.remove(id);
     }
 }
